requirements.txt
flask
flask-cors
gunicorn
instaloader
ultralytics
easyocr
pillow
